Clean up external links when printing
To improve printed pages' usability, we appended link URLs to link
text in parentheses. Unfortunately, our external links point to
an interstitial page that is long and ugly when printed.

This commit saves the original external URL as a data attribute on
the anchor element for later retrieval by our print stylesheet.

https://github.local/CFGOV/platform/issues/3812